Move ExecutionManagerTestBuilder under MockDescriptors
